Update scripts inventory, diagrams & chembl schemas
Refresh scripts inventory manifest and catalog to reflect a large reorganization of scripts (many CI helpers moved/renamed into scripts/data and new scripts/diagrams tooling). Update references, counts and generated_at timestamp in configs/quality/scripts_inventory_manifest.json and adjust scripts/catalog.yaml and sonar-project.properties accordingly. Apply related code/schema tweaks in src/bioetl (chembl activity/silver schema changes) and update test expectations/snapshots (architecture and unit tests, plus chembl_activity_schema.json snapshot) to match the new paths and contract changes.

+++ tests/unit/infrastructure/schemas/test_silver_pipeline_contracts.py
@@ -45,7 +45,7 @@ PIPELINE_SCHEMA_EXPECTATIONS: dict[str, list[tuple[str, pa.DataType]]] = {
         ("ligand_efficiency_le", pa.float64()),
         ("ligand_efficiency_lle", pa.float64()),
         ("ligand_efficiency_sei", pa.float64()),
-        ("manual_curation_flag", pa.int64()),
+        ("manual_curation_flag", pa.float64()),  # Float for nullable int
         ("molecule_id", pa.string()),
         ("molecule_pref_name", pa.string()),
         ("original_activity_id", pa.float64()),  # Float for nullable int
